+++ src/main/java/com/cedarsoftware/util/io/JsonWriter.java
@@ -1,7 +1,5 @@
 package com.cedarsoftware.util.io;
 
-import static com.cedarsoftware.util.io.JsonObject.ITEMS;
-
 import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
 import java.io.Flushable;
@@ -30,8 +28,9 @@ import java.util.Set;
 
 import com.cedarsoftware.util.reflect.Accessor;
 import com.cedarsoftware.util.reflect.ClassDescriptors;
-
 import lombok.Getter;
+
+import static com.cedarsoftware.util.io.JsonObject.ITEMS;
 
 /**
  * Output a Java object graph in JSON format.  This code handles cyclic
@@ -79,6 +78,8 @@ public class JsonWriter implements WriterContext, Closeable, Flushable
     private static final Object[] byteStrings = new Object[256];
     private static final String NEW_LINE = System.getProperty("line.separator");
     private static final Long ZERO = 0L;
+    @Deprecated
+    public static final String TYPE = "TYPE";
 
     @Getter
     private final WriteOptions writeOptions;
@@ -2019,5 +2020,60 @@ public class JsonWriter implements WriterContext, Closeable, Flushable
         }
         Long id = this.objsReferenced.get(o);
         return id == null ? null : Long.toString(id);
+    }
+
+    /**
+     * Write out special characters "\b, \f, \t, \n, \r", as such, backslash as \\
+     * quote as \" and values less than an ASCII space (20hex) as "\\u00xx" format,
+     * characters in the range of ASCII space to a '~' as ASCII, and anything higher in UTF-8.
+     *
+     * @param s String to be written in UTF-8 format on the output stream.
+     * @param output Writer to which the UTF-8 string will be written to
+     * @throws IOException if an error occurs writing to the output stream.
+     */
+    public static void writeJsonUtf8String(String s, final Writer output) throws IOException
+    {
+        output.write('\"');
+        final int len = s.length();
+
+        for (int i = 0; i < len; i++)
+        {
+            char c = s.charAt(i);
+
+            if (c < ' ')
+            {    // Anything less than ASCII space, write either in \\u00xx form, or the special \t, \n, etc. form
+                switch (c)
+                {
+                    case '\b':
+                        output.write("\\b");
+                        break;
+                    case '\f':
+                        output.write("\\f");
+                        break;
+                    case '\n':
+                        output.write("\\n");
+                        break;
+                    case '\r':
+                        output.write("\\r");
+                        break;
+                    case '\t':
+                        output.write("\\t");
+                        break;
+                    default:
+                        output.write(String.format("\\u%04X", (int)c));
+                        break;
+                }
+            }
+            else if (c == '\\' || c == '"')
+            {
+                output.write('\\');
+                output.write(c);
+            }
+            else
+            {   // Anything else - write in UTF-8 form (multi-byte encoded) (OutputStreamWriter is UTF-8)
+                output.write(c);
+            }
+        }
+        output.write('\"');
     }
 }